WIP Reworking Resource class
-- Reworking Resource class: only managed by ResourceManager (private
new, delete). Change it for a template class in the aim to encapsulate
any type of data: each data type stored in its own typed container (in
the same way that LoaderManager class). This avoids collision with
unique ID (for example Texture can have the ID 1 while SimTaDynMap has
also the IS 1).
-- WIP: ILoader interface is no longer working loadFromFile will produce
a segfault and need to be reworked.
-- Rename ILoader.hpp to ILoader.tpp
-- Rename IResource to Resource.tpp
-- LoaderManager.tpp: add method for creating all loaders.
-- Adding a loader for OpenGL textures.
-- SimTaDynMap, ResourceTests: refactore code because of the Resource
class refactorisation.